Map Trello cards result to class
