+++ templates/index.html
@@ -22,14 +22,13 @@
           {% for toDoItem in toDoItems %}
           <tr>
             <td>
-              {{toDoItem.name}}
+              {{toDoItem.title}}
             </td>
             <td>
               {{ toDoItem.status }}
               {% if toDoItem.status == 'Completed' %}
-              Complete ✔
+              ✔
               {% else %}
-              Incomplete 
               <a class="btn btn-light" href="{{ url_for('complete_item', id=toDoItem.id) }}" title="Complete item">✔</a>
               {% endif %}
             </td>
