--- trello_config_reader.py
@@ -1,5 +1,0 @@
-def get_config(filename):
-    file = open(f'trello_config/{filename}.txt')
-    config = file.read()
-    file.close()
-    return config
